+++ lib/src/text_analyzer/_index.dart
@@ -5,5 +5,8 @@
 export 'english_analyzer/english.dart';
 export 'latin_language_analyzer/latin_language_analyzer.dart';
 export 'term_co_occurrence_graph.dart';
+export 'n_gram_range.dart';
 export 'text_analysis_extensions.dart';
 export 'text_analyzer.dart';
+export 'token.dart';
+export 'tokenizing_strategy.dart';