+++ lib/src/text_analyzer/latin_language_analyzer/_keyword_extractor.dart
@@ -1,0 +1,5 @@
+// BSD 3-Clause License
+// Copyright ©2022, GM Consult Pty Ltd
+// All rights reserved
+
+part of 'latin_language_analyzer.dart';