+++ lib/src/text_analyzer/tokenizing_strategy.dart
@@ -1,7 +1,7 @@
 // BSD 3-Clause License
 // Copyright ©2022, GM Consult Pty Ltd
 
-import 'package:text_analysis/src/text_tokenizer/_index.dart';
+import 'package:text_analysis/src/_index.dart';
 
 /// The strategy to apply when splitting text to [Token]s.
 enum TokenizingStrategy {