+++ lib/src/text_tokenizer/_index.dart
@@ -2,10 +2,9 @@
 // Copyright ©2022, GM Consult Pty Ltd
 // All rights reserved
 
-export 'n_gram_range.dart';
+
 // export 'text_analysis_extensions.dart';
 // export 'text_analyzer.dart';
 // export 'text_document.dart';
 export 'text_tokenizer.dart';
-export 'token.dart';
-export 'tokenizer_strategy.dart';
+